requirements.txt
celery==5.4.0
redis==5.0.8
fastapi==0.112.2
uvicorn==0.30.6
pymongo~=4.8.0
pydantic~=2.8.2
requests~=2.32.3
pyjwt~=2.9.0